Move playstore related functions to a separate utils so that we can mock easily in tests
